implement validation
