+++ pkg/graphql/request.go
@@ -5,14 +5,24 @@ import (
 	"errors"
 	"io"
 	"io/ioutil"
+
+	"github.com/jensneuse/graphql-go-tools/pkg/ast"
+	"github.com/jensneuse/graphql-go-tools/pkg/astparser"
+	"github.com/jensneuse/graphql-go-tools/pkg/astvalidation"
+	"github.com/jensneuse/graphql-go-tools/pkg/operationreport"
 )
 
-var ErrEmptyRequest = errors.New("the provided request is empty")
+var (
+	ErrEmptyRequest = errors.New("the provided request is empty")
+	ErrNilSchema    = errors.New("provided schema is nil")
+)
 
 type Request struct {
 	OperationName string          `json:"operation_name"`
 	Variables     json.RawMessage `json:"variables"`
 	Query         string          `json:"query"`
+
+	document ast.Document
 }
 
 func UnmarshalRequest(reader io.Reader, request *Request) error {
@@ -28,8 +38,20 @@ func UnmarshalRequest(reader io.Reader, request *Request) error {
 	return json.Unmarshal(requestBytes, &request)
 }
 
-func (r Request) ValidateForSchema(schema *Schema) (valid bool, errors OperationValidationErrors) {
-	return true, nil
+func (r *Request) ValidateForSchema(schema *Schema) (result ValidationResult, err error) {
+	if schema == nil {
+		return ValidationResult{Valid: false, Errors: nil}, ErrNilSchema
+	}
+
+	var report operationreport.Report
+	r.document, report = astparser.ParseGraphqlDocumentString(r.Query)
+	if report.HasErrors() {
+		return operationValidationResultFromReport(report)
+	}
+
+	validator := astvalidation.DefaultOperationValidator()
+	validator.Validate(&r.document, &schema.document, &report)
+	return operationValidationResultFromReport(report)
 }
 
 func (r *Request) Normalize(schema *Schema) error {
