+++ pkg/graphql/request_test.go
@@ -5,6 +5,9 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/jensneuse/graphql-go-tools/pkg/starwars"
 )
 
 func TestUnmarshalRequest(t *testing.T) {
@@ -29,5 +32,55 @@ func TestUnmarshalRequest(t *testing.T) {
 		assert.NoError(t, err)
 		assert.Equal(t, "Hello", request.OperationName)
 		assert.Equal(t, "query Hello { hello }", request.Query)
+	})
+}
+
+func TestRequest_ValidateForSchema(t *testing.T) {
+	t.Run("should return error when schema is nil", func(t *testing.T) {
+		request := Request{
+			OperationName: "Hello",
+			Variables:     nil,
+			Query:         `query Hello { hello }`,
+		}
+
+		result, err := request.ValidateForSchema(nil)
+		assert.Error(t, err)
+		assert.Equal(t, ErrNilSchema, err)
+		assert.Equal(t, ValidationResult{Valid: false, Errors: nil}, result)
+	})
+
+	t.Run("should return gql errors when validation fails", func(t *testing.T) {
+		request := Request{
+			OperationName: "Goodbye",
+			Variables:     nil,
+			Query:         `query Goodbye { goodbye }`,
+		}
+
+		schema, err := NewSchemaFromString("schema { query: Query } type Query { hello: String }")
+		require.NoError(t, err)
+
+		result, err := request.ValidateForSchema(schema)
+		assert.NoError(t, err)
+		assert.False(t, result.Valid)
+		assert.Greater(t, result.Errors.Count(), 0)
+	})
+
+	t.Run("should return valid result when validation is successful", func(t *testing.T) {
+		starwars.SetRelativePathToStarWarsPackage("../starwars")
+		schemaBytes := starwars.Schema(t)
+
+		schema, err := NewSchemaFromString(string(schemaBytes))
+		require.NoError(t, err)
+
+		rawRequest := starwars.LoadQuery(t, starwars.FileSimpleHeroQuery, nil)
+
+		var request Request
+		err = UnmarshalRequest(bytes.NewBuffer(rawRequest), &request)
+		require.NoError(t, err)
+
+		result, err := request.ValidateForSchema(schema)
+		assert.NoError(t, err)
+		assert.True(t, result.Valid)
+		assert.Nil(t, result.Errors)
 	})
 }
